Added Update Box (see description!)
Found a way to send multiple parameters to .php file but note this is not the normal way (the 2 different normal ways are commented out in the same section as the 'Update4User' method). Must work on a way to use the normal ways as they are probably more commonplace.